+++ CartMS/Models/ProductModel.cs
@@ -7,9 +7,9 @@ namespace CartMS.Models
     [PrimaryKey(nameof(seller_id), nameof(product_id))]
     public class ProductModel
     {
-        public long seller_id { get; set; }
+        public int seller_id { get; set; }
 
-        public long product_id { get; set; }
+        public int product_id { get; set; }
 
         public string name { get; set; } = "";
 
@@ -19,9 +19,9 @@ namespace CartMS.Models
 
         public string description { get; set; } = "";
 
-        public decimal price { get; set; }
+        public float price { get; set; }
 
-        public decimal freight_value { get; set; }
+        public float freight_value { get; set; }
 
         public DateTime created_at { get; set; }
