+++ Common/Entities/StockItem.cs
@@ -2,9 +2,9 @@
 {
 	public class StockItem
 	{
-        public long seller_id { get; set; }
+        public int seller_id { get; set; }
 
-        public long product_id { get; set; }
+        public int product_id { get; set; }
 
         public int qty_available { get; set; }
