+++ Common/Events/ShipmentNotification.cs
@@ -4,8 +4,8 @@ namespace Common.Events
 {
     public record ShipmentNotification
     (
-        long customerId,
-        long orderId,
+        int customerId,
+        int orderId,
         DateTime eventDate,
         int instanceId,
         ShipmentStatus status = ShipmentStatus.approved
