+++ CustomerMS/Repositories/ICustomerRepository.cs
@@ -6,7 +6,7 @@ namespace CustomerMS.Repositories
 {
 	public interface ICustomerRepository
 	{
-        CustomerModel? GetById(long customerId);
+        CustomerModel? GetById(int customerId);
         CustomerModel Insert(CustomerModel customer);
         CustomerModel Update(CustomerModel customer);
     }
